Support separate compilation
Inline trees can now be read form TASTY. However, positions are not set
correctly. This remains to be implemented.
