+++ tests/run/inline/Test_2.scala
@@ -1,0 +1,18 @@
+object Test {
+
+  import p.inlines._
+
+  def main(args: Array[String]): Unit = {
+    println(f(10))
+    println(f(f(10)))
+
+    track("hello") { println("") }
+
+    val o = new Outer
+    val i = new o.Inner
+    println(i.m)
+    //println(i.g)
+    //println(i.h)
+  }
+
+}
