+++ tests/run/inlinePower/Test_2.scala
@@ -1,0 +1,9 @@
+import p.pow.power
+object Test {
+
+  def main(args: Array[String]): Unit = {
+    println(power(2.0, 10))
+    def x = 2.0
+    println(power(x, 11))
+  }
+}
