+++ tests/run/inlinePower/power_1.scala
@@ -1,0 +1,13 @@
+package p
+
+object pow {
+
+  @dotty.annotation.inline
+  def power(x: Double, n: Int): Double =
+    if (n == 0) 1.0
+    else if (n == 1) x
+    else {
+      val y = power(x, n / 2)
+      if (n % 2 == 0) y * y else y * y * x
+    }
+}
